Add combine() to return an xarray dataset with combined lat/lon coords
Ideally, I think this should be accomplished with xr.combine_by_coords,
but this is not available in the current conda release, and
xr.auto_combine doesn't work (or is extremely slow) and is being
deprecated. As a workaround, xr.open_mfdataset() accomplishes what we
need--a single combined dataset, including all towers.

An additional step invoking the "garbage collector" tries to free up
memory, but it's not straightforward to verify if this actually frees
up the memory or not.