Fix multiindex sorting in to_dataframe()
Simply resetting/setting the multiindices does not in general correctly
sort the indices; need to explicitly call sort_index(). There is a
slight performance penalty to resorting O(1) second, BUT the resulting
call to pd.DataFrame.to_xarray() may slow down from O(0.1) s to as much
as O(10) s if the indices are not properly sorted.