Fix rounding bug with default wrf.Tower.to_dataframe()
For large mesoscale timesteps >= 1 s, previous behavior (without a
specified time-step size, default) would round timestamps to the nearest
second and deal with any single-to-double precision conversion issues--
providing a clean dataframe index as a result.

However, for smaller timesteps that are fractions of a second, this
results in dataframes with non-unique indices. The new rounding is to
the nearest millisecond. With 6 decimal places in the tsout towers, the
actual output precision is 1e-6 h or 0.0036 s ~ O(1e-3) s