Remove eta_levels from requirements/add submit_all
Eta levels do not need to be specified - you can let WRF do that
on its own if you give the number of levels.

Added scripts that will be in the main directory to submit all
wps, real, and wrf submission scripts. For running with several
cases, this is extremely useful to just kick off all at once.