Bug fix for dx,dy and add MERRA2 to IC/BCs
dx and dy were being calculated incorrectly from the given dx,dy
and the parent_grid_ratio. This has been fixed using np.prod() so
that the correct dx,dy are calculated.

MERRA2 currently has to be downloaded manually... adjusted the
setup so that it doesn't download anything, links the manually
downloaded files (need to be in icbc_directory/) in the submit_wps
script, and does not run ungrib.