Bug fix - dimension / coordinate issue resolved
Most of the tests were with xr.Datasets converted to pd.dataframe
before being fed into the calc_spectra code. I didn't realize that
when you convert xr.Dataset().to_dataframe().to_xarray() that you
do not get the original xr.Dataset back... dimension and
coordinate changes occur and errors were thrown. The fix makes it
so that when a xr.Dataset is read in that each dimension has a
coordinate and any coordinates without a dimension become variables