Updating TRI and VRM calculations for footprint
Using a footprint in the generic filter allows for interesting
techniques in calculating TRI / VRM. In this case, we wanted to
sub-sample the terrain as it was interpolated (smoothed) and the
TRI / VRM values were very low compared to the native terrain
data. This footprint allows for sampling a window in only some
cells and not others.