Adding class to calculate eta levels
Class will allow for user-specified heights to be converted to eta
levels (given some other parameters), if the heights do not reach
the model top, then the code can fill in levels (using the same
formulation that WRF uses), and if there are any jumps in d(eta)
the user can smooth the levels. Smoothing will change the eta
levels, so if there is constant spacing at some point, the spline
may adjust it so that it is not quite constant anymore.

Alternatively, if the user has eta levels already but wants to
smooth out any kinks, the code accepts eta levels and then the
user can simply call the 'smooth_eta_levels function to generate a
smoothed list.